kuflow-samples-kubot-desktop-screenshot/kuflow_samples_kubot_desktop_screenshot/tasks-desktop-screenshot.py
#
# MIT License
#
# Copyright (c) 2022 KuFlow
#
# Permission is hereby granted, free of charge, to any person obtaining a copy
# of this software and associated documentation files (the "Software"), to deal
# in the Software without restriction, including without limitation the rights
# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
# copies of the Software, and to permit persons to whom the Software is
# furnished to do so, subject to the following conditions:
#
# The above copyright notice and this permission notice shall be included in all
# copies or substantial portions of the Software.
#
# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
# SOFTWARE.
#

import os

from PIL import ImageGrab


def take_a_desktop_screenshot_to_home():
    # Get the path to the user's home directory
    home_dir = os.path.expanduser("~")

    # Output file path
    image_path = os.path.join(home_dir, "desktop-screenshot.png")

    # Task screenshot
    screenshot = ImageGrab.grab()
    screenshot.save(image_path)

    print(f"Desktop screenshot has been written in {image_path}")


if __name__ == "__main__":
    take_a_desktop_screenshot_to_home()
