kuflow-samples-kubot-desktop-screenshot/pyproject.toml
[tool.poetry]
name = "kuflow-samples-kubot-desktop-screenshot"
version = "0.1.0"
description = "KuFlow KuBot example. Take a user desktop screenshot"
authors = ["KuFlow S.L. <kuflow@kuflow.com>"]
homepage = "https://kuflow.com/"
documentation = "https://docs.kuflow.com/"
repository = "https://github.com/kuflow/kuflow-samples-python"
license = "MIT"
readme = "README.md"
packages = [{include = "kuflow_samples_kubot_desktop_screenshot"}]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "~3.11"
kuflow-rest = "^2.0.4"
pillow = "^10.4.0"

[tool.poetry.group.dev.dependencies]
mypy = "^1.3.0"
ruff = ">=0.1.8,<1.0.0"
pytest = "^7.4.2"

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
testpaths = ["tests"]
python_files= "test_*.py"
